+++ src/oci/key_management/__init__.py
@@ -7,6 +7,8 @@
 from __future__ import absolute_import
 
 
+from .ekm_client import EkmClient
+from .ekm_client_composite_operations import EkmClientCompositeOperations
 from .kms_crypto_client import KmsCryptoClient
 from .kms_crypto_client_composite_operations import KmsCryptoClientCompositeOperations
 from .kms_management_client import KmsManagementClient
@@ -15,4 +17,4 @@ from .kms_vault_client import KmsVaultClient
 from .kms_vault_client_composite_operations import KmsVaultClientCompositeOperations
 from . import models
 
-__all__ = ["KmsCryptoClient", "KmsCryptoClientCompositeOperations", "KmsManagementClient", "KmsManagementClientCompositeOperations", "KmsVaultClient", "KmsVaultClientCompositeOperations", "models"]
+__all__ = ["EkmClient", "EkmClientCompositeOperations", "KmsCryptoClient", "KmsCryptoClientCompositeOperations", "KmsManagementClient", "KmsManagementClientCompositeOperations", "KmsVaultClient", "KmsVaultClientCompositeOperations", "models"]